FIX guard against any user destroying relations

+++ src/sandbox/App2/server/api/relation.ts
@@ -66,14 +66,19 @@ export class RelationAPI<
     return item.values;
   }
 
-  async destroyRelationsForItem({
-    id,
-  }: MutationDestroyItemArgs): Promise<number> {
-    return this.store.Relation.destroy({
-      where: {
-        [Op.or]: [{ itemId1: id }, { itemId2: id }],
-      },
-    });
+  async destroyRelationsForItem(
+    { id }: MutationDestroyItemArgs,
+    userId: UserId
+  ): Promise<number> {
+    if (userId === this.userId) {
+      return this.store.Relation.destroy({
+        where: {
+          [Op.or]: [{ itemId1: id }, { itemId2: id }],
+        },
+      });
+    }
+
+    return 0;
   }
 
   async getAllRelatedItems({
